Update AWSI test levels and automation tests
