--- AutomatedTesting/Gem/PythonTests/AWS/Windows/aws_metrics/__init__.py
@@ -1,6 +1,0 @@
-"""
-Copyright (c) Contributors to the Open 3D Engine Project.
-For complete copyright and license terms please see the LICENSE at the root of this distribution.
-
-SPDX-License-Identifier: Apache-2.0 OR MIT
-"""
