+++ AutomatedTesting/Gem/PythonTests/AWS/client_auth/__init__.py
