+++ AutomatedTesting/Gem/PythonTests/AWS/core/__init__.py
@@ -4,4 +4,3 @@ For complete copyright and license terms please see the LICENSE at the root of t
 
 SPDX-License-Identifier: Apache-2.0 OR MIT
 """
-
