--- AutomatedTesting/Levels/AWS/ClientAuth/ClientAuth.ly
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:19f2c4454bb395cdc0a36d1e45e6a384bbd23037af1a2fb93e088ecfa0f10e5b
-size 9343
