--- AutomatedTesting/Levels/AWS/ClientAuth/filelist.xml
@@ -1,6 +1,0 @@
-<download name="ClientAuth" type="Map">
- <index src="filelist.xml" dest="filelist.xml"/>
- <files>
-  <file src="level.pak" dest="level.pak" size="E1A" md5="fecbc160ebc2186184504482e3e2eba3"/>
- </files>
-</download>
