--- AutomatedTesting/Levels/AWS/ClientAuthPasswordSignIn/level.pak
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f318a1787069385de291660f79e350cea2ca2c3ef3b5e0576686066bd9c49395
-size 3667
