--- AutomatedTesting/Levels/AWS/ClientAuthPasswordSignUp/filelist.xml
@@ -1,6 +1,0 @@
-<download name="ClientAuthPasswordSignUp" type="Map">
- <index src="filelist.xml" dest="filelist.xml"/>
- <files>
-  <file src="level.pak" dest="level.pak" size="E44" md5="acfaa325178533f135cd683d3782cc8e"/>
- </files>
-</download>
