--- AutomatedTesting/Levels/AWS/ClientAuthPasswordSignUp/level.pak
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:87882b64688a77815d93c6973929fa21b89dc6c13d4866c710124ce2cd0f411e
-size 3652
