--- AutomatedTesting/Levels/AWS/Metrics/Metrics.ly
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:023992998ab5a1d64b38dacd1d5e1a9dc930ff704289c0656ed6eaba6951d660
-size 9066
